Adding base classes
